Set the controls window in fullscreen to be slightly opaque

+++ src/support.c
@@ -3131,7 +3131,7 @@ gboolean detect_volume_option()
         av[ac++] = g_strdup_printf("%s", mplayer_bin);
     }
     av[ac++] = g_strdup_printf("-volume");
-    av[ac++] = g_strdup_printf("%i",idledata->volume);
+    av[ac++] = g_strdup_printf("%i",(gint)idledata->volume);
     av[ac++] = g_strdup_printf("-noidle");
     av[ac++] = g_strdup_printf("-softvol");
     av[ac] = NULL;
